fargate_jobs/google_scraper/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
aiosqlite==0.21.0
alphashape==1.3.1
annotated-types==0.7.0
anyio==4.9.0
apify_client==1.12.0
apify_shared==1.4.1
attrs==25.3.0
Automat==25.4.16
aws-encryption-sdk==3.3.0
aws-encryption-sdk-cli==4.2.0
base64io==1.0.3
beautifulsoup4==4.13.4
boto3==1.38.42
botocore==1.38.42
Brotli==1.1.0
certifi==2025.7.9
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
click-log==0.4.0
colorama==0.4.6
constantly==23.10.4
Crawl4AI==0.7.0
cryptography==45.0.4
cssselect==1.3.0
decorator==5.2.1
distro==1.9.0
fake-http-header==0.3.5
fake-useragent==2.2.0
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.7.0
googlesearch-python==1.3.0
greenlet==3.2.3
groq==0.30.0
h11==0.16.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.4
humanize==4.12.3
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.7.0
incremental==24.7.2
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lark==1.2.2
litellm==1.74.3
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.6.3
networkx==3.5
nltk==3.9.1
numpy==2.3.1
openai==1.96.0
outcome==1.3.0.post0
packaging==25.0
pillow==11.3.0
playwright==1.53.0
pluggy==1.6.0
propcache==0.3.2
psutil==7.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.2
pyOpenSSL==25.1.0
pyperclip==1.9.0
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-tornasync==0.6.0.post2
pytest-trio==0.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
rank-bm25==0.2.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rpds-py==0.26.0
rtree==1.4.0
s3transfer==0.13.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
sentence-transformers==5.0.0
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.7
sympy==1.14.0
tf-playwright-stealth==1.2.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.53.2
trimesh==4.7.0
trio==0.30.0
Twisted==25.5.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
wheel==0.45.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zope.interface==7.2
